style touchup for model info, changed visualization settings icon to funnel, fixed naming of model-info
