+++ frontend/src/features/model-info/components/model-info-header.tsx
