+++ frontend/src/features/model-info/components/model-info.tsx
@@ -1,6 +1,6 @@
 import { ModelData } from "~/features/model-ranking/types";
-import { ModelInfoChart } from "./mode-info-chart";
-import { ModelExample, ModelInfoExamples } from "./mode-info-exmaples";
+import { ModelInfoChart } from "./model-info-chart";
+import { ModelExample, ModelInfoExamples } from "./model-info-examples";
 import { ModelInfoDetails } from "./model-info-details";
 import { ModelInfoHeader } from "./model-info-header";
 
@@ -25,12 +25,12 @@ export function ModelInfo({ model }: ModelInfoProps) {
   return (
     <div className="space-y-8 flex-col p-4 rounded-xl border-2">
       <ModelInfoHeader name={model.model_name} />
-      <div className="p-4 space-y-10 lg:grid lg:grid-cols-4 gap-4 ">
-        <div className="flex flex-col gap-6 col-span-2">
+      <div className="space-y-10 lg:grid lg:grid-cols-4 gap-4">
+        <div className="flex flex-col gap-6 col-span-2 m-0">
           <ModelInfoDetails model={model} />
           <ModelInfoChart model={model} />
         </div>
-        <div className="col-span-2">
+        <div className="flex col-span-2">
           <ModelInfoExamples examples={examples} />
         </div>
       </div>
